refactor(command-visualizer): extract model/ scope (shared pure contract)
Move the shared, pure, HTML-free contract into model/: frame-state,
columns, overlays, timeline. Add model/geometry.ts extracting the Pos and
Range interfaces plus the orderRange helper out of the old serialize.ts so
both logic/ and render/ depend on geometry without either importing the
other. Repoint intra-model and data/ imports.

Pure relocation + import rewrite; no behavior change.
