+++ packages/command-visualizer/src/logic/chain.ts
