+++ packages/command-visualizer/src/logic/fixture-extract.ts
