+++ packages/command-visualizer/src/logic/hat-allocator.test.ts
