+++ packages/command-visualizer/src/logic/hat-allocator.ts
