+++ packages/command-visualizer/src/logic/pipeline.ts
