+++ packages/command-visualizer/src/logic/tokenize.ts
