+++ packages/command-visualizer/src/model/columns.ts
