+++ packages/command-visualizer/src/model/timeline.ts
