+++ packages/command-visualizer/src/render/serialize.ts
