+++ packages/command-visualizer/src/render/svg-wrap.ts
