+++ packages/command-visualizer/src/render/symbols.ts
